feat: add tests

+++ src/components/SongItem/SongItem.tsx
@@ -1,24 +1,39 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { useDispatch } from 'react-redux';
 import AddIcon from '@material-ui/icons/Add';
 import RemoveIcon from '@material-ui/icons/Remove';
 import Song from '../../types/Song'
 import { Container, SongTitle, AddedBy, VoteCount, VoteButtons, DeleteIcon } from './styles'
 import { playlistsAsyncActions } from '../../store/slices/playlists/slice';
+import Confirmation from '../Confirmation';
+import { useSelector } from 'react-redux';
+import { selectCurrentPlaylist } from '../../store/slices/playlists/selectors';
+import { selectCurrentUser } from '../../store/slices/authentication/selectors';
 
 type SongProps = {
     song: Song
 }
 
-
 const SongItem = ({ song }: SongProps) => {
+    const currentUser = useSelector(selectCurrentUser);
+    const currentPlaylist = useSelector(selectCurrentPlaylist)
+    const isOwner = currentUser?.name === currentPlaylist?.ownerName;
+    const [confirmationIsVisible, setConfirmationIsVisible] = useState(false)
     const dispatch = useDispatch()
     
-    const deleteSong = () => {
-        dispatch(playlistsAsyncActions.deleteSong(song.id))
+    const confirmAction = (isConfirmed: boolean) => {
+        if (isConfirmed) {
+            dispatch(playlistsAsyncActions.deleteSong(song.id))
+        }
+        setConfirmationIsVisible(false)
+    }
+
+    const deleteOnClick = () => {
+        setConfirmationIsVisible(true)
     }
 
     return (
+        <>
         <Container>
             <SongTitle>
                 {song.title}
@@ -33,8 +48,10 @@ const SongItem = ({ song }: SongProps) => {
                 <AddIcon />
                 <RemoveIcon />
             </VoteButtons>
-            <DeleteIcon onClick={deleteSong}/>
+            {isOwner && <DeleteIcon onClick={deleteOnClick} data-testid="delete-icon"/>}
         </Container>
+        {confirmationIsVisible && <Confirmation message={"delete this song"} confirm={confirmAction} data-testid="confirmation-container"/>}
+        </>
     )
 }
